[DEMO] Align demo drivers and MEs for 64 bits

Closes #56

See merge request smartobject/soo!36

+++ agency/linux/soo/include/soo/dcm/datacomm.h
@@ -26,6 +26,6 @@ void datacomm_init(void);
 
 bool datacomm_ready_to_send(void);
 
-void datacomm_send(void *ME_buffer, size_t size, uint32_t prio);
+void datacomm_send(void *ME_buffer, uint32_t size, uint32_t prio);
 
 #endif /* DATACOMM_H */
